Show of label of PlantYpe prediction

+++ lib/infrastructure/models/plant-from-gcloud.dart
@@ -5,7 +5,7 @@ class PlantFromGCloud {
   });
 
   final String plantClass;
-  final int confidence;
+  final double confidence;
 
   factory PlantFromGCloud.fromJson(Map<String, dynamic> json) =>
       PlantFromGCloud(
